move accessibility tests into a separate class
