+++ code/test/Templates.Test/Templates.Test.csproj
@@ -98,6 +98,7 @@
     <Compile Include="FileEncodingTests.cs" />
     <Compile Include="CodeStyleEnforcementTests.cs" />
     <Compile Include="GenTests\LanguageComparisonTests.cs" />
+    <Compile Include="VisualTests\AccessibilityTests.cs" />
     <Compile Include="WACK\Wpf\WindowsAppCertKitTests.cs" />
     <None Include=".codeclonesettings" />
     <None Include="App.config" />
